requirements.txt
streamlit==1.41.1
python-dotenv==1.0.1

langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.3.0

chromadb==0.5.23

pypdf==5.1.0
tiktoken==0.8.0

pandas==2.2.3
matplotlib==3.10.0

sentence-transformers==3.3.1
torch==2.5.1
